Refactor some of the image methods
